Moved 404 page to screen folder (semantics) Cleaned up message component for long-text
