+++ client/src/components/Message.js
@@ -1,10 +1,11 @@
 import React from "react";
-import { MessageUser } from "./styled";
+import { MessageUser, MessageText } from "./styled";
 
 const Message = ({ text, username, createdAt, self }) => {
   return (
     <div title={new Date(createdAt).toLocaleString()}>
-      <MessageUser self={self}>{username}</MessageUser>: {text}
+      <MessageUser self={self}>{username}</MessageUser>:{" "}
+      <MessageText>{text}</MessageText>
     </div>
   );
 };
